Check if the service is enabled and restarted

+++ install.sh
@@ -47,13 +47,13 @@ main() {
 
 	cprintf blue "\nRestarting Waybar..."
 
-	if systemctl --user is-enabled waybar.service &> /dev/null; then
-		systemctl --user restart waybar.service
-	else
-		pkill waybar
-		waybar &> /dev/null &
-		disown
+	if ! (systemctl --user is-enabled waybar.service &&
+		systemctl --user restart waybar.service) &>/dev/null; then
+    	pkill waybar
+    	waybar &>/dev/null &
+    	disown
 	fi
+
 
 	if ((ERRORS > 0)); then
 		cprintf red "\nInstallation completed with $ERRORS errors"
